Renamed package parser to http.
