+++ src/main/java/io/github/sebastianschmidt/galop/GalopModule.java
@@ -4,7 +4,7 @@ import com.google.inject.AbstractModule;
 import io.github.sebastianschmidt.galop.administration.AdministrationModule;
 import io.github.sebastianschmidt.galop.commons.CommonsModule;
 import io.github.sebastianschmidt.galop.configuration.ConfigurationModule;
-import io.github.sebastianschmidt.galop.parser.ParserModule;
+import io.github.sebastianschmidt.galop.http.HttpModule;
 import io.github.sebastianschmidt.galop.proxy.ProxyModule;
 
 final class GalopModule extends AbstractModule {
@@ -13,7 +13,7 @@ final class GalopModule extends AbstractModule {
     protected void configure() {
         install(new CommonsModule());
         install(new ConfigurationModule());
-        install(new ParserModule());
+        install(new HttpModule());
         install(new ProxyModule());
         install(new AdministrationModule());
         bind(Starter.class);
