+++ src/main/java/io/github/sebastianschmidt/galop/http/UnsupportedTransferEncodingException.java
@@ -1,4 +1,4 @@
-package io.github.sebastianschmidt.galop.parser;
+package io.github.sebastianschmidt.galop.http;
 
 import java.io.IOException;
